// src/admin-app/src/app/orders/constants/order-columns.enum.ts
export enum OrderColumn {
  sgOrderId = 'sgOrderId',
  createDate = 'createDate',
  modifyDate = 'modifyDate',
  company = 'company',
  paymentType = 'paymentType',
  paymentStatus = 'paymentStatus',
  price = 'price',
  currency = 'currency',
  invoices = 'invoices',
  profaInvoices = 'profaInvoices',
  author = 'author',
  note = 'note',
  status = 'status',
  actions = 'actions'
}